requirements.txt
numpy~=1.26.2
pyserial~=3.5
pandas~=2.1.3
plotly~=5.18.0